Add argument for number of days to keep

+++ sykepic/predict/sync.py
@@ -21,6 +21,7 @@ def main(args):
 
 def download(args):
     s3 = boto3.resource('s3', endpoint_url='https://a3s.fi')
+    num_downloads = 0
     for file in list(allas.ls(args.allas, s3=s3)):
         date = ifcb.sample_to_datetime(file)
         day_path = date.strftime('%Y/%m/%d')
@@ -31,11 +32,12 @@ def download(args):
         if not save_dir.is_dir():
             save_dir.mkdir(parents=True)
         try:
-            print(f'[INFO] Downloading {file}')
+            num_downloads += 1
             allas.download(Path(args.allas)/file, save_dir, s3)
         except Exception as e:
             print(f'[ERROR] {file}: {e}', file=stderr)
             continue
+    print(f'[INFO] Downloaded {num_downloads} new files')
 
 
 def upload(args):
@@ -43,8 +45,9 @@ def upload(args):
     s3 = boto3.resource('s3', endpoint_url='https://a3s.fi')
     today = datetime.date.today()
     # today = datetime.date(2018, 7, 4)
-    day_dirs = []
     # Find all directories corresponding to one day
+    day_dirs = []
+    # find_head_dirs() modifies day_dirs in place
     find_head_dirs(local, day_dirs)
     for day_dir in day_dirs:
         try:
@@ -71,22 +74,24 @@ def upload(args):
                     zipped_files.append(file)
         print(f'[INFO] Uploading to {allas_zip.parent}')
         allas.upload(local_zip, allas_zip, s3)
-        if args.remove:
-            shutil.rmtree(day_dir)
-            print(f'[INFO] Removed {day_dir}')
-            if args.remove > 1:
-                # Remove local zip file
-                local_zip.unlink()
-                # Remove month directory if it's empty
-                try:
-                    day_dir.parent.rmdir()
-                except OSError:
-                    pass
-                print(f'[INFO] Removed {local_zip}')
-        if args.clean_allas:
-            print(f'[INFO] Cleaning {args.clean_allas} in Allas')
-            for file in zipped_files:
-                allas.delete(f'{args.clean_allas}/{file.name}', s3)
+        # Removing local copies
+        if (date - today).days() > args.keep:
+            if args.remove:
+                shutil.rmtree(day_dir)
+                print(f'[INFO] Removed {day_dir}')
+                if args.remove > 1:
+                    # Remove local zip file
+                    local_zip.unlink()
+                    # Remove month directory if it's empty
+                    try:
+                        day_dir.parent.rmdir()
+                    except OSError:
+                        pass
+                    print(f'[INFO] Removed {local_zip}')
+            if args.clean_allas:
+                print(f'[INFO] Cleaning {args.clean_allas} in Allas')
+                for file in zipped_files:
+                    allas.delete(f'{args.clean_allas}/{file.name}', s3)
 
 
 def find_head_dirs(path, head_dirs=[]):
